Second pass at editting the interval values - pass the scoreboard id or scoreboard type to load the scoreboard in the viewmodel - the time transformer needs to confirm to and from the time data for this screen to work     - time data just needs ints for the values instead of longs - use 2 text fields to edit the minute and second -

+++ app/src/main/java/com/dgnt/quickScoreboardCreator/domain/scoreboard/model/time/TimeData.kt
@@ -1,7 +1,7 @@
 package com.dgnt.quickScoreboardCreator.domain.scoreboard.model.time
 
 data class TimeData(
-    val minute: Long,
-    val second: Long,
-    val centiSecond: Long,
+    val minute: Int,
+    val second: Int,
+    val centiSecond: Int,
 )
